refactor(stats): extract shared Apex chart options builder for Stimulus charts
Move analysis chart rendering options into a reusable module so existing
and generic analysis charts share the same styling and stacking behaviour.
